feat: 인증/인가에 해당하지 않는 API의 디렉토리 위치 변경
